fixe project setup and add the function of login

+++ back_end.flutter/src/main/java/com/example/back_end/flutter/Reposetry/EmploiReposetry.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.back_end.flutter.models.Emploi_du_temp;
 
-public interface EmploiReposetry extends JpaRepository<Emploi_du_temp, Double>{
+public interface EmploiReposetry extends JpaRepository<Emploi_du_temp, Long>{
 
 }
