+++ back_end.flutter/src/main/java/com/example/back_end/flutter/Reposetry/ProfesseurReposetry.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.back_end.flutter.models.*;
 
-public interface ProfesseurReposetry extends JpaRepository<professeur, Double> {
+public interface ProfesseurReposetry extends JpaRepository<professeur, Long> {
 
 }
